app/src/test/java/com/fasten/executor_driver/presentation/vehicleoptions/VehicleOptionsViewStatePendingTest.java
package com.fasten.executor_driver.presentation.vehicleoptions;

import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.verifyNoMoreInteractions;

import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mock;
import org.mockito.junit.MockitoJUnitRunner;

@RunWith(MockitoJUnitRunner.class)
public class VehicleOptionsViewStatePendingTest {

  private VehicleOptionsViewStatePending viewState;

  @Mock
  private VehicleOptionsViewActions codeViewActions;

  @Before
  public void setUp() {
    viewState = new VehicleOptionsViewStatePending();
  }

  @Test
  public void testActions() {
    // Действие:
    viewState.apply(codeViewActions);

    // Результат:
    verify(codeViewActions).enableReadyButton(false);
    verify(codeViewActions).showVehicleOptionsList(true);
    verify(codeViewActions).showVehicleOptionsPending(true);
    verify(codeViewActions).showVehicleOptionsListErrorMessage(false);
    verifyNoMoreInteractions(codeViewActions);
  }
}